==> code/02-config.R <==
# config

# dependencies -----------------------------------------------------------

# load packages
source("code/01-load-packages.R")

# set seed for reproducibility -------------------------------------------

set.seed(42)

# plot config ------------------------------------------------------------

# define theme for plots
theme_set(theme_classic())

# colorblind friendly palette: 2 colors for populations
pal_pop <- met.brewer("OKeeffe1")[c(4, 8)]
names(pal_pop) <- c("Bon", "Cai")

# colorblind friendly palette: 2 colors for treatments
pal_treat <- moma.colors("Avedon")[c(6, 9)]
names(pal_treat) <- c(
  "Control",
  "Herbivore-induced"
)
# make "Control" a bit darker so it's more visible in the figures
pal_treat["Control"] <-
  colorspace::darken(pal_treat["Control"], amount = 0.05) # 5% darker

# figure dimensions & resolution
fig_width <- 6.5 # width in inches
fig_height <- 4.5 # height in inches
fig_dpi <- 640 # resolution in dpi

# create directories: figures and tables
if (!dir.exists("figures")) {
  dir.create("figures")
}
if (!dir.exists("tables")) {
  dir.create("tables")
}
